requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-embed-video==1.4.2
djangorestframework==3.13.1
idna==3.3
install==1.3.5
Pillow==9.0.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
